requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-cognitiveservices-speech==1.37.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imbalanced-learn==0.12.0
imblearn==0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
multidict==6.0.5
numpy==1.26.3
openai==0.28.0
pandas==2.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyspark==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.12.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
threadpoolctl==3.2.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.0
watchdog==3.0.0
websockets==12.0
Werkzeug==3.0.1
xgboost==2.0.3
yarl==1.9.4
zipp==3.17.0
